Fix issue with non-required options - Laravel 4
The problem occurs when you add a @SWG\Parameter that has the "required = false" but the Swagger frontend shows that it is a required field and it is not possible to execute "Try it" button without adding something in non-required field. This fix will solve the problem and it will be possible to execute the request without filling the non-required fields.